pyproject.toml
[project]
name = "sresrabrownies"
version = "0.1.0"
description = ""
authors = [
    {name = "Thiago Pinotti",email = "thiago.pinotti@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "streamlit (>=1.50.0,<2.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "markdown (>=3.9,<4.0)",
    "weasyprint (>=66.0,<67.0)",
    "num2words (>=0.5.14,<0.6.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
